+++ extension/lib/reduxConstants.js
@@ -1,3 +1,6 @@
+export const ADD = 'ADD';
+export const CHA = 'CHA';
 export const DEL = 'DEL';
 export const NEW = 'NEW';
+export const REM = 'REM';
 export const SET = 'SET';